change the decimal sql type and change some entities properties to int

+++ backend/Netatmo.Dashboard.Api/Models/Device.cs
@@ -21,7 +21,7 @@ namespace Netatmo.Dashboard.Api.Models
     {
         public int RfStatus { get; set; }
         public int BatteryVp { get; set; }
-        public decimal BatteryPercent { get; set; }
+        public int BatteryPercent { get; set; }
         public ModuleDeviceType Type { get; set; }
     }
 }
